Revert "some ordering and cart related changes, not finalized yet"
